Truncate dir and file names that exceed the Windows maximum name length of 255
Directory and file names that exceed the Windows maximum file name length of 255 characters are now automatically truncated.

You can use the following sample album URLs to test this behavior:

https://play.qobuz.com/album/0843436060947
https://play.qobuz.com/album/mttqwiwzz3ccb

For readers: This 255-character limit applies specifically to directory and file names (including file extension). Do not confuse it with MAX_PATH constant, which is the 260-character path length limitation for applications that are not long-path aware, which is not the case for this application.