ProgressBar for downloads,. Also, various additional improvements and fixes.
Added incremental progress reporting for downloads, through a ProgressBar. Also, various additional improvements and fixes.

Note: This PR does NOT depend on any other PR. In other words, it is completely independent to the main branch.

This PR introduces progress reporting for downloads in the application. A ProgressBar has been added to the UI, and all relevant download operations (albums, tracks, playlists, artists, and user favorites) now reports the overall progress.

The main changes were written to the switch (linkType) in qbdlxForm.cs class, and the download functions in DownloadTrack.cs and DownloadAlbum.cs classes.

Related to these changes, getLinkType function was made asynchronous, renamed to getLinkTypeAsync, and adapted asynchronous calls to this function from other parts of the source-code.

Also, this PR it includes the following improvements and fixes:

1. Added a themeable custom progressbar user control (see: CustomProgressBar.cs class)

Note: This class was almost entirely generated by ChatGPT.

2. Implemented theming logic for the custom progressbar in Theming.cs class:

3. Fixed tab-order for the entire application (main form and login form), and set TabStop property to false for minimize and maximize buttons.

Note: having loads of "hidden" controls in the designer doesn't help, so I've written partial logic of the tab-order in qbdlxForm_Load event handler:

4. Made the main form double buffered to prevent flickering (specially in the ProgressBar control):

5. The "GET" button and the associated textbox are now temporarily disabled while a download operation is running:

This avoids awkward behavior in the application if the user click the button twice while a download is running.

6. SendURL method now sends a click to the "GET" button instead of calling directly the getLinkType function:

5. The "GET" buttons in the search panel are also temporarily disabled while the main "GET" button is disabled / a download operation is running:

This allows the "GET" button to behave the way explained in point number 5.

7. Fixed all the compiler suggestions/informational messages, among them are: string interpolation simplification, type instantiation simplification, removal of unused parameters, making variables read-only, and deleting (commented) unused public fields.

Just eevrything was fixed, except the suggestions to fix naming rule violations, in order to respect the author's programming habits.

8. Album title label is now automatically truncated when it exceeds the main form bounds:

And I think I'm not forgetting anything else.

------------------

**This will possibly be my last PR to this repository, as C# is not my main language and all of this took me many hours to get it into perfect working condition.

I really appreciate that you created this application and made it open-source. And I hope all my PRs so far can be accepted, as I consider them necessary and/or useful little improvements.

Best regards!**

PS: I've tested many Qobuz urls from artists, playlists, labels and tracks, and didn't found errors nor unexpected behaviors related to the changes that I've applied in this and previous PRs. However, I insist that C# is not my main language, so I may have made some silly mistake somewhere without realizing it.